move app/components/features into kanbanwave
