--- src/app/components/features/index.ts
@@ -1,5 +1,0 @@
-export { default as CardDraggable } from './CardDraggable';
-export { default as CardDroppable } from './CardDroppable';
-export { default as Card } from './Card';
-export { default as List } from './List';
-export { default as StrictModeDroppable } from './StrictModeDroppable';
