+++ src/app/components/index.ts
@@ -1,3 +1,2 @@
 export * from './common';
-export * from './features';
 export * from './layout';
