+++ src/app/pages/Board.tsx
@@ -1,13 +1,12 @@
 import {
   Title,
-  List,
-  StrictModeDroppable
 } from 'app/components';
 import { useCallback } from 'react';
 import * as Dummy from 'app/dummy';
-import { AddItemForm, KWItemType, useKanbanCard, useKanbanList } from 'kanbanwave';
+import { AddItemForm, KWItemType, List, useKanbanCard, useKanbanList } from 'kanbanwave';
 import { DragDropContext, DropResult } from 'react-beautiful-dnd';
 import { useLocation } from 'react-router-dom';
+import StrictModeDroppable from 'kanbanwave/StrictModeDroppable';
 
 const Board = () => {
   const location = useLocation();
