+++ src/kanbanwave/List.tsx
@@ -1,13 +1,15 @@
-import { AddItemForm, KWList, useKanbanCard } from 'kanbanwave';
+import AddItemForm from './AddItemForm';
+import { useKanbanCard } from './KanbanStorageProvider';
+import { KWList } from './types';
 import {
   IconButton,
   Subtitle,
-  CardDroppable,
 } from 'app/components';
 import { useCallback } from 'react';
-import { Card } from 'app/components';
 import * as Dummy from 'app/dummy';
 import { Draggable } from 'react-beautiful-dnd';
+import CardDroppable from './CardDroppable';
+import Card from './Card';
 
 type ListProps = {
   index: number;
