+++ src/kanbanwave/StrictModeDroppable.tsx
