+++ src/kanbanwave/index.ts
@@ -10,3 +10,5 @@ export {
 } from './KanbanStorageProvider';
 
 export { default as AddItemForm } from './AddItemForm'
+export { default as Card } from './Card'
+export { default as List } from './List'
